# openapi (3).yaml
openapi: 3.0.1
info:
  title: "Bank"
  description: ""
  version: "1"
servers:
- url: https://virtserver.swaggerhub.com/3_heroes/Laba_5/1
...
paths:
...
  /user/{username}:
    get:
      tags:
        - Users
      summary: Get user by user name
      operationId: getUserByName
      parameters:
        - name: username
          in: path
          description: The name that needs to be fetched.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
      security:
        - User_api_key: []
    put:
      tags:
        - Users
      summary: Updated user
      description: This can only be done by the logged in user.
      operationId: updateUser
      parameters:
        - name: username
          in: path
          description: name that need to be updated
          required: true
          schema:
            type: string
      responses:
        '400':
          description: Invalid user supplied
        '404':
          description: User not found
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        description: Updated user object
        required: true
      security:
        - User_api_key: []
    delete:
      tags:
        - Users
      summary: Delete user
      description: This can only be done by the logged in user.
      operationId: deleteUser
      parameters:
        - name: username
          in: path
          description: The name that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
      security:
        - User_api_key: []
...
components:
  schemas:
...
    User:
      required:
      - password
      - username
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        password:
          type: string
          format: password
        ClientName:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        status:
          type: string
          description: User Status
          enum:
            - user
            - manager
...
  securitySchemes:
    User_api_key:
      type: apiKey
      name: User_api_key
      in: header
    Manager_api_key:
      type: apiKey
      name: Manager_api_key
      in: header